design & refactor: ui변경 적용 및 컴포넌트 분리
