+++ src/components/Layout/Layout.style.js
@@ -4,83 +4,8 @@ import { Link } from "react-router-dom";
 //메인 컨테이너 스타일
 export const Container = styled.div`
   max-width: 1140px;
-  min-height: 90vh;
   margin: 0 auto;
-  padding: 120px 0;
-`;
-
-/**------헤더 스타일------- */
-export const HeaerParentBox = styled.div`
-  position: fixed;
-  z-index: 10;
-  width: 100%;
-  background-color: #f9f7f7;
-  border-bottom: 1px solid #dbe2ef;
-`;
-
-export const HeaderBox = styled.header`
-  position: relative;
-  z-index: 10;
-  max-width: 1400px;
-  height: 80px;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-//슬로건 박스
-export const SloganBox = styled.div`
-  p {
-    color: #112d4e;
-    line-height: 40px;
-  }
-`;
-
-//코드 프로젝트 메뉴 박스
-export const CodeProjectMenuBox = styled.div`
-  ul{
-    height: 40px;
-    display: flex;
-    align-items: center;
-    
-  }
-
-  a {
-    padding: 0 8px;
-    font-weight: bold;
-    font-size: 1rem;
-    line-height: 20px;
-  }
-
-  a:first-child {
-    border-right: 1px solid #dbe2ef;
-  }
-`
-
-//로그인 버튼
-export const LogineButton = styled.button`
-  width: 80px;
-  height: 40px;
-  padding: 10px;
-  background-color: #3f72af;
-  color: #fff;
-  border: none;
-  border-radius: 12px;
-  cursor: pointer;
-`;
-//로고
-export const Logo = styled.h1`
-  color: #112d4e;
-  font-size: 2.5rem;
-  font-weight: bold;
-  font-family: "FjallaOne";
-`;
-
-//링크 스타일
-export const LogoLink = styled(Link)`
-  cursor: pointer;
-  color: #112d4e;
+  padding: 2.75rem 0;
 `;
 
 //로그인 된 유저 박스
@@ -90,10 +15,12 @@ export const UserLoginBox = styled.div`
   gap: 12px;
   justify-content: center;
   align-items: center;
+
   img {
     cursor: pointer;
   }
 `;
+
 //프로필 이미지
 export const ProfileImage = styled.img`
   width: 40px;
@@ -116,28 +43,30 @@ export const MenuBox = styled.div`
 `;
 
 //메뉴 리스트 박스
-
 export const MenuListBox = styled.ul`
   height: 100%;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+
   li {
     width: 100%;
-    flex:1;
+    flex: 1;
     text-align: center;
     line-height: 40px;
     list-style: none;
   }
+
   li:hover {
-    background-color: #D4DBE9; /* 마우스를 올렸을 때 배경 색상 변경 */
+    background-color: #d4dbe9; /* 마우스를 올렸을 때 배경 색상 변경 */
   }
+
   li:last-child {
     border-top: 1px solid #fff;
   }
 
-  .logout{
+  .logout {
     cursor: pointer;
   }
 `;
@@ -158,6 +87,7 @@ export const FooterBox = styled.div`
   background-color: #dbe2ef;
   height: 160px;
 `;
+
 export const DevelopersBox = styled.div`
   display: flex;
   color: #112d4e;
@@ -166,14 +96,17 @@ export const DevelopersBox = styled.div`
     font-weight: bold;
   }
 `;
+
 export const CorporationBox = styled.div`
   display: flex;
   color: #112d4e;
   gap: 8px;
   cursor: pointer;
+
   div {
     padding: 0 8px;
   }
+
   div:not(:last-child) {
     border-right: 1px solid #f9f7f7; //고객센터 박스만 제외하고 오른쪽에 테두리
   }
